# backend/app/src/main/java/com/ssafy/d205/global/exception/GlobalExceptionHandler.java
package com.ssafy.d205.global.exception;

import org.springframework.dao.DataIntegrityViolationException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.FieldError;
import org.springframework.context.MessageSourceResolvable;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.http.converter.HttpMessageNotReadableException;
import org.springframework.web.bind.MissingRequestHeaderException;
import org.springframework.web.bind.MissingServletRequestParameterException;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;
import org.springframework.web.method.annotation.HandlerMethodValidationException;

import java.util.stream.Collectors;

/**
 * 예외를 HTTP 응답으로 옮깁니다.
 *
 * <p>이게 없으면 스프링 기본 처리가 나가는데, 그건 예외 종류를 대체로 500으로 뭉개거나
 * 응답 본문에 내부 정보를 흘립니다. 잘못된 요청과 서버 고장을 클라이언트가 구분할 수
 * 있어야 재시도 여부를 판단할 수 있습니다.
 */
@RestControllerAdvice
public class GlobalExceptionHandler {

    @ExceptionHandler(MethodArgumentNotValidException.class)
    public ResponseEntity<ErrorResponse> handleValidation(MethodArgumentNotValidException e) {
        String message = e.getBindingResult().getFieldErrors().stream()
                .map(FieldError::getDefaultMessage)
                .collect(Collectors.joining(" "));
        return ResponseEntity.badRequest().body(new ErrorResponse("INVALID_REQUEST", message));
    }

    /**
     * 쿼리 파라미터나 경로 변수의 제약 위반입니다.
     *
     * <p>요청 본문(@Valid @RequestBody)은 MethodArgumentNotValidException 으로 오지만,
     * 파라미터에 직접 붙인 제약은 스프링 프레임워크 6.1부터 이 예외로 옵니다. 둘을
     * 같은 코드로 내보내는 것은 클라이언트에게 "요청이 규칙에 안 맞는다"는 사실이
     * 같기 때문입니다. 어디가 틀렸는지는 메시지가 알려줍니다.
     */
    @ExceptionHandler(HandlerMethodValidationException.class)
    public ResponseEntity<ErrorResponse> handleParameterValidation(HandlerMethodValidationException e) {
        String message = e.getAllErrors().stream()
                .map(MessageSourceResolvable::getDefaultMessage)
                .collect(Collectors.joining(" "));
        return ResponseEntity.badRequest().body(new ErrorResponse("INVALID_REQUEST", message));
    }

    /**
     * 본문을 아예 읽지 못한 경우입니다. JSON 이 깨졌거나, 값이 필드의 타입에 맞지 않습니다.
     *
     * <p>신고 사유처럼 열거형으로 받는 값이 목록에 없으면 여기로 옵니다. @Valid 가
     * 도는 것은 본문이 객체로 바뀐 뒤라, 바뀌지 못한 요청은 검증까지 가지 못합니다.
     * 이 핸들러가 없으면 스프링 기본 응답이 나가고 클라이언트는 code 로 분기할 수
     * 없습니다.
     *
     * <p><b>예외 메시지를 그대로 쓰지 않습니다.</b> 잭슨의 메시지에는 클래스 이름과
     * 필드 경로가 들어 있어서 내부 구조가 그대로 흘러나갑니다.
     */
    @ExceptionHandler(HttpMessageNotReadableException.class)
    public ResponseEntity<ErrorResponse> handleUnreadableBody(HttpMessageNotReadableException e) {
        return ResponseEntity.badRequest()
                .body(new ErrorResponse("INVALID_REQUEST", "요청 본문의 값이 형식에 맞지 않습니다."));
    }


    /** 필수 쿼리 값(분석 탭 positions 의 matchId 등)이 빠진 경우입니다. 스프링 기본 400 에는 code 가 없습니다. */
    @ExceptionHandler(MissingServletRequestParameterException.class)
    public ResponseEntity<ErrorResponse> handleMissingParameter(MissingServletRequestParameterException e) {
        return ResponseEntity.badRequest()
                .body(new ErrorResponse("INVALID_REQUEST", "필수 값 " + e.getParameterName() + " 이 없습니다."));
    }

    @ExceptionHandler(MissingRequestHeaderException.class)
    public ResponseEntity<ErrorResponse> handleMissingHeader(MissingRequestHeaderException e) {
        return ResponseEntity.badRequest()
                .body(new ErrorResponse("MISSING_HEADER", e.getHeaderName() + " 헤더가 필요합니다."));
    }

    @ExceptionHandler(UnknownCallerException.class)
    public ResponseEntity<ErrorResponse> handleAccountNotFound(UnknownCallerException e) {
        return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(new ErrorResponse("ACCOUNT_NOT_FOUND", "계정을 찾을 수 없습니다."));
    }

    /**
     * X-User-Id 는 왔는데 그 계정의 토큰이 없거나 맞지 않습니다. 정지 검사보다 앞에서
     * 나므로 토큰 없는 요청은 SUSPENDED 를 볼 수 없습니다.
     *
     * <p>관리자 로그인이 없을 때 SecurityConfig 가 내는 401 과 코드가 같습니다. 뜻도 같아서
     * (누구인지 증명하지 못했다) 일부러 맞췄습니다.
     */
    @ExceptionHandler(UnauthorizedException.class)
    public ResponseEntity<ErrorResponse> handleUnauthorized(UnauthorizedException e) {
        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                .body(new ErrorResponse("UNAUTHORIZED", "계정 토큰이 없거나 맞지 않습니다."));
    }

    /**
     * 정지된 계정입니다. 계정 발급과 X-User-Id 가 붙은 모든 요청에서 나옵니다.
     *
     * <p>ACCOUNT_NOT_FOUND 와 코드를 구분합니다. 그쪽은 클라이언트가 계정 발급을 다시
     * 불러야 하는 상황이고, 이쪽은 다시 불러도 같은 자리입니다 - 발급도 막혀 있습니다.
     *
     * <p>사유를 담지 않는 이유는 {@link SuspendedAccountException} 에 있습니다.
     */
    @ExceptionHandler(SuspendedAccountException.class)
    public ResponseEntity<ErrorResponse> handleSuspended(SuspendedAccountException e) {
        return ResponseEntity.status(HttpStatus.FORBIDDEN)
                .body(new ErrorResponse("SUSPENDED", "정지된 계정입니다."));
    }

    /**
     * 상대를 찾을 수 없습니다. 부르는 사람이 없는 경우(ACCOUNT_NOT_FOUND)와 코드를
     * 구분합니다. 전자는 클라이언트가 계정 발급을 다시 불러야 하고, 후자는 "그
     * 사용자가 없습니다"를 보여주면 됩니다.
     *
     * <p>차단 관계일 때도 이 코드입니다. 차단당했다는 사실을 알려주면 차단한 사람이
     * 드러나므로, 없는 사용자와 구분되지 않게 두는 것이 의도입니다.
     */
    @ExceptionHandler(TargetUserNotFoundException.class)
    public ResponseEntity<ErrorResponse> handleTargetNotFound(TargetUserNotFoundException e) {
        return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(new ErrorResponse("TARGET_NOT_FOUND", "상대를 찾을 수 없습니다."));
    }

    @ExceptionHandler(SelfFriendRequestException.class)
    public ResponseEntity<ErrorResponse> handleSelfRequest(SelfFriendRequestException e) {
        return ResponseEntity.badRequest()
                .body(new ErrorResponse("SELF_FRIEND_REQUEST", "자기 자신에게 친구 요청을 보낼 수 없습니다."));
    }


    @ExceptionHandler(AlreadyFriendsException.class)
    public ResponseEntity<ErrorResponse> handleAlreadyFriends(AlreadyFriendsException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("ALREADY_FRIENDS", "이미 친구입니다."));
    }

    @ExceptionHandler(FriendRequestAlreadySentException.class)
    public ResponseEntity<ErrorResponse> handleRequestAlreadySent(FriendRequestAlreadySentException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("REQUEST_ALREADY_SENT", "이미 보낸 친구 요청입니다."));
    }

    @ExceptionHandler(FriendRequestNotFoundException.class)
    public ResponseEntity<ErrorResponse> handleRequestNotFound(FriendRequestNotFoundException e) {
        return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(new ErrorResponse("FRIEND_REQUEST_NOT_FOUND", "친구 요청을 찾을 수 없습니다."));
    }

    @ExceptionHandler(NotFriendsException.class)
    public ResponseEntity<ErrorResponse> handleNotFriends(NotFriendsException e) {
        return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(new ErrorResponse("NOT_FRIENDS", "친구가 아닙니다."));
    }

    /**
     * 초대할 상대가 로비나 경기 중입니다. 초대 토스트는 홈에서만 뜨므로 지금 보내도 상대가
     * 볼 수 없고 3분 뒤 조용히 만료됩니다. 보낸 사람에게 그 사실을 바로 알립니다.
     */
    @ExceptionHandler(TargetInGameException.class)
    public ResponseEntity<ErrorResponse> handleTargetInGame(TargetInGameException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("TARGET_IN_GAME", "게임 중인 친구에게는 초대를 보낼 수 없습니다."));
    }

    @ExceptionHandler(NicknameTakenException.class)
    public ResponseEntity<ErrorResponse> handleNicknameTaken(NicknameTakenException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("NICKNAME_TAKEN", "이미 사용 중인 닉네임입니다."));
    }

    /**
     * 닉네임에 금칙어가 있습니다 (S15P21D205-1017). 400 인 이유는 입력을 고치면 되는 문제이기 때문이고,
     * INVALID_REQUEST 와 코드를 나눈 이유는 화면이 "글자 규칙" 과 "쓸 수 없는 말" 을 다르게 안내해야
     * 하기 때문입니다. 어느 말에 걸렸는지는 담지 않습니다 - 담으면 목록을 알아내는 데 쓰입니다.
     */
    @ExceptionHandler(NicknameForbiddenException.class)
    public ResponseEntity<ErrorResponse> handleNicknameForbidden(NicknameForbiddenException e) {
        return ResponseEntity.badRequest()
                .body(new ErrorResponse("NICKNAME_FORBIDDEN", "쓸 수 없는 닉네임입니다."));
    }

    /** 같은 경기에서 같은 사람을 이미 신고했습니다 (S15P21D205-1017). 화면은 "이미 신고했습니다" 로 안내합니다. */
    @ExceptionHandler(ReportAlreadySentException.class)
    public ResponseEntity<ErrorResponse> handleReportAlreadySent(ReportAlreadySentException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("REPORT_ALREADY_SENT", "이 경기에서 그 사람을 이미 신고했습니다."));
    }

    /**
     * 닉네임 변경이 uk_users_nickname에 걸린 경우입니다.
     *
     * <p>서비스가 미리 조회해 확인하지만 그 사이에 다른 요청이 같은 닉네임을 차지할 수
     * 있습니다. 그 경쟁에서 진 요청이 여기로 옵니다. 사용자에게는 위와 같은 상황이므로
     * 같은 409를 줍니다.
     */
    @ExceptionHandler(DataIntegrityViolationException.class)
    public ResponseEntity<ErrorResponse> handleConflict(DataIntegrityViolationException e) {
        return ResponseEntity.status(HttpStatus.CONFLICT)
                .body(new ErrorResponse("CONFLICT", "요청이 다른 데이터와 충돌했습니다. 다시 시도해 주세요."));
    }

    @ExceptionHandler(NicknameGenerationFailedException.class)
    public ResponseEntity<ErrorResponse> handleNicknameGeneration(NicknameGenerationFailedException e) {
        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                .body(new ErrorResponse("NICKNAME_GENERATION_FAILED", "계정 발급에 실패했습니다. 다시 시도해 주세요."));
    }
}