# backend/docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminAnalyticsTable" : {
        "properties" : {
          "columns" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "rows" : {
            "items" : {
              "items" : { },
              "type" : "array"
            },
            "type" : "array"
          },
          "unavailable" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "기기 식별자 기반 계정, 유저 검색, 친구, 차단, 접속 상태 API.\n\nX-User-Id 는 인증이 아니라 식별입니다. 되돌릴 수 없는 연산만 자격증명(X-Device-Id)을 함께 요구합니다. 오류 코드 표와 하트비트 규칙은 docs/client-guide.md 를 보세요.",
    "title" : "D205 API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/analytics/positions" : {
      "get" : {
        "operationId" : "positions",
        "parameters" : [
          {
            "in" : "query",
            "name" : "matchId",
            "required" : true,
            "schema" : {
              "maxLength" : 64,
              "minLength" : 0,
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminAnalyticsTable"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-analytics-controller"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "description" : "운영",
      "url" : "https://j15d205.p.ssafy.io"
    },
    {
      "description" : "로컬",
      "url" : "http://localhost:8080"
    }
  ]
}